fix(conway)!: bound committee terms at ratification and widen expiry epochs to uint64 (#2222)
* fix(conway): enforce committee term limits


* fix(conway): harden committee term validation


* fix(ledger)!: widen committee expiry epochs to uint64

UpdateCommitteeGovAction.CredEpochs held absolute expiry epochs as uint.
The CDDL encodes that map value as `epoch = uint .size 8`, so on a 32-bit
build a conforming update_committee action could not carry an expiry at or
above 2^32. CommitteeMember.ExpiryEpoch, which stores the enacted form of
the same quantity, is already uint64; epoch_interval parameters such as
CommitteeTermLimit stay uint64 and MaxEpoch stays uint.

CredEpochs and the NewUpdateCommitteeGovAction parameter become
map[*Credential]uint64, dropping the widening conversions in
validateCommitteeTerms and ToPlutusData. This also lets the committee term
overflow case assert a full uint64 expiry on GOARCH=386.


* fix(conway): make the committee term bound a ratification predicate

validateCommitteeTerms ran from UtxoValidateGovActionWellFormedness, which
is registered as a UTxO validation rule in conway/rules.go and reused by
dijkstra/rules.go, so an over-limit UpdateCommittee expiry invalidated the
transaction and any block carrying it.

In cardano-ledger the bound is not a transaction-validity check.
validCommitteeTerm is evaluated in ratifyTransition (Ratify.hs:338) and
gates enactment; actionWellFormed (Gov.hs:393), the GOV analogue of this
function, does not mention committee terms. A transaction proposing an
expiry beyond the maximum term is accepted and simply not ratified, so
rejecting it here would reject blocks the Haskell node accepts. The verdict
also depended on the validating node's current epoch rather than on the
transaction, so mempool admission, block re-validation and replay could
disagree, and the rejection was permanent where the reference bound loosens
each epoch.

The predicate is exported as ValidateCommitteeTerm(action, pp, currentEpoch)
for an epoch-boundary caller and the UTxO call site is removed. Taking the
epoch as a parameter drops the common.CurrentEpochState capability and
CurrentEpochStateUnavailableError, so no ledger state has to implement a new
method to keep proposing committee updates, and the conformance harness no
longer needs a state-provider wrapper.

The bound is now currentEpoch + maxTermLength computed directly. The
reference addEpochInterval is EpochNo (n + fromIntegral m) on Word64 and
wraps; the previous saturating form accepted expiries near 2^64 that the
reference rejects.


---------

+++ ledger/conway/rules_gov_transition_test.go
@@ -268,7 +268,7 @@ func TestUtxoValidateGovActionWellFormedness(t *testing.T) {
 		}
 		action := &common.UpdateCommitteeGovAction{
 			Credentials: []common.Credential{cred},
-			CredEpochs: map[*common.Credential]uint{
+			CredEpochs: map[*common.Credential]uint64{
 				&cred: 500,
 			},
 		}
@@ -290,7 +290,7 @@ func TestUtxoValidateGovActionWellFormedness(t *testing.T) {
 		}
 		action := &common.UpdateCommitteeGovAction{
 			Credentials: []common.Credential{removeCred},
-			CredEpochs: map[*common.Credential]uint{
+			CredEpochs: map[*common.Credential]uint64{
 				&addCred: 500,
 			},
 		}
@@ -341,7 +341,7 @@ func TestUtxoValidateGovActionWellFormedness(t *testing.T) {
 			}
 			action := &common.UpdateCommitteeGovAction{
 				Credentials: []common.Credential{credA, credB},
-				CredEpochs: map[*common.Credential]uint{
+				CredEpochs: map[*common.Credential]uint64{
 					&credA: 500,
 					&credB: 600,
 				},
